backend/requirement.txt
Flask
opencv-python
numpy
matplotlib
flask_cors
ultralytics 
pytest
python-dotenv